DRR - Cpptraj: ClusterMatrix V0->V1: Enable saving of sieved ClusterMatrix; only non-ignored pairs are saved, as well as an array describing which frames are ignored. Also store #elements and #rows as unsigned 8 byte integers for very large array support.

+++ src/TriangleMatrix.h
@@ -11,11 +11,11 @@
 class TriangleMatrix : public DataSet {
   public :
     int Nrows()                const { return (int)nrows_;            }
-    int Nelements()            const { return (int)nelements_;        }
+    size_t Nelements()         const { return nelements_;             }
     double operator[](int idx) const { return (double)elements_[idx]; }
 
     TriangleMatrix();
-    TriangleMatrix(int);
+    TriangleMatrix(size_t);
     TriangleMatrix(const TriangleMatrix&);
     ~TriangleMatrix();
     TriangleMatrix & operator=(const TriangleMatrix &);
